% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoteFunction.R
\name{initSSH}
\alias{initSSH}
\title{Initiate SSH}
\usage{
initSSH(login, password = NULL, pwfile = NULL, port = 22,
  PS1 = "remFunPrompt>>>", regexPS1 = NULL, tempdir.loc = NULL,
  tempdir.rem = NULL, Rscript = NULL, timeout.con = 20,
  timeout.cmd = 20, verbosity = 1, delay = c(1, 1, 1, 5, 5, 5),
  delay2 = delay * 1.5, use.exist = TRUE)
}
\arguments{
\item{login}{login in the form user@host}

\item{password}{password for the ssh connection}

\item{pwfile}{file whose first line contains password (overrides \code{password})}

\item{tempdir.loc}{directory on local machine to store files}

\item{tempdir.rem}{directory on remote machine to store files}

\item{Rscript}{path to the Rscript executable on the remote machine}

\item{timeout.con}{duration after which a connection attempt is considered as timed out}

\item{timeout.cmd}{duration after which a bash command is considered timed out}

\item{verbosity}{verbosity of the status messages}

\item{delay}{waiting time between connection attempts (lower limit)}

\item{delay2}{waiting time between connection attempts (upper limit)}

\item{use.exist}{should functions and results already available in \code{tempdir.loc} be used}
}
\value{
A list with functions to define and execute R functions, see \code{details}.
}
\description{
Initiates the SSH connection and returns an object
to define and execute functions remotely
}
\details{
The returned list contains the following functions:
\tabular{ll}{
\code{createRemoteFunction(fun, fun.name=NULL,
cache=TRUE, transfer=TRUE, timeout=NULL,
show.output=TRUE)}\tab
Defines a function \code{fun} with name \code{fun.name}.
Returns a function with the same arguments as \code{fun} that remotely evaluates \code{fun}.
The flag \code{cache} determines whether inputs and results should be cached on the local machine.
The maximal allowed duration of execution in seconds is specified by \code{timeout}.
The flag \code{show.output} determines whether output to \code{stdout} on remote machine should be
mirrored on the local machine. If \code{transfer=FALSE}, the function is not uploaded automatically
and function \code{updateAndTransfer} has to be called before execution. \cr\cr
\code{updateAndTransfer} \tab Batch uploads all functions created by \code{createRemoteFunction}.
If the directory \code{tempdir.loc} was also used in previous sessions, then also functions defined
in these previous sessions will be uploaded.\cr\cr
\code{uploadPackage(package)}\tab Uploads a package to the remote machine\cr\cr
\code{removePackage(package)}\tab Removes an uploaded package from the remote machine\cr\cr
\code{uploadGlobalSource(text=NULL,file=NULL)}\tab Uploads R code either given as file or as character string
that is sourced before every function execution.
If both \code{text} and \code{file} are specified, \code{text} will be used.\cr\cr
\code{uploadGlobal(...)}\tab Uploads objects to the remote machine. Syntax is the same as for \code{save}.\cr\cr
\code{removeGlobal()}\tab Remove all global objects and sources from the remote machine.\cr\cr\cr
\code{getFun(name)}\tab returns the source code of the function with the specified \code{name}\cr\cr
\code{listFuns()}\tab returns a character vector with the names of all registered functions.\cr\cr
\code{execFun(name)}\tab returns the function wrapper to execute the respective function remotely.
If \code{fun(arg1, arg2, ...)}, then the function can be called by \code{execFun("fun")(arg1,arg2,...)}\cr\cr
\code{getFunDataDir(name)} \tab returns the data directory of the function named \code{name} on the local machine\cr\cr
\code{getRemFunDataDir(name)} \tab returns the data directory of the function named \code{name} on the remote machine\cr\cr
\code{uploadFunData(name,files=NULL)} \tab uploads files from the local to the remote data directory\cr\cr
\code{downloadFunData(name,files=NULL)} \tab downloads files from the remote to the local data directory\cr\cr
\code{clearFunData(name,files=NULL,del.rem=TRUE, del.loc=TRUE)} \tab deletes data from the local and/or remote machine\cr\cr
\code{closeCon()}\tab Closes the connection and cleans up files (but does not delete functions and their inputs and outputs)
}

The result list contains beside these functions also the objects \code{ssh} and \code{rsync}.
These objects are the result of calls to \code{initInteractiveSSH} and \code{initRsync} of the packages
\code{interactiveSSH} and \code{rsyncFacility}.
}
